+++ modules/src/ics02_client/events.rs
@@ -7,50 +7,50 @@ use anomaly::BoxError;
 
 use crate::ics02_client::height::Height;
 use serde_derive::{Deserialize, Serialize};
-use std::collections::HashSet;
 use std::convert::{TryFrom, TryInto};
 use tendermint::block;
 
 /// The content of the `type` field for the event that a chain produces upon executing the create client transaction.
-pub const CREATE_EVENT_TYPE: &str = "create_client";
-pub const UPDATE_EVENT_TYPE: &str = "update_client";
+const CREATE_EVENT_TYPE: &str = "create_client";
+const UPDATE_EVENT_TYPE: &str = "update_client";
 
 /// The content of the `key` field for the attribute containing the client identifier.
-pub const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
+const CLIENT_ID_ATTRIBUTE_KEY: &str = "client_id";
 
 /// The content of the `key` field for the attribute containing the client type.
-pub const CLIENT_TYPE_ATTRIBUTE_KEY: &str = "client_type";
+const CLIENT_TYPE_ATTRIBUTE_KEY: &str = "client_type";
 
 /// The content of the `key` field for the attribute containing the height.
-pub const CONSENSUS_HEIGHT_ATTRIBUTE_KEY: &str = "consensus_height";
+const CONSENSUS_HEIGHT_ATTRIBUTE_KEY: &str = "consensus_height";
 
-/// A list of all the event `type`s that this module is capable of parsing
-fn event_types() -> HashSet<String> {
-    vec![CREATE_EVENT_TYPE.to_string(), UPDATE_EVENT_TYPE.to_string()]
-        .into_iter()
-        .collect()
+pub fn try_from_tx(event: &tendermint::abci::Event) -> Option<IBCEvent> {
+    match event.type_str.as_ref() {
+        CREATE_EVENT_TYPE => Some(IBCEvent::CreateClient(CreateClient(
+            extract_attributes_from_tx(event),
+        ))),
+        UPDATE_EVENT_TYPE => Some(IBCEvent::UpdateClient(UpdateClient(
+            extract_attributes_from_tx(event),
+        ))),
+        _ => None,
+    }
 }
 
-pub fn try_from_tx(event: tendermint::abci::Event) -> Option<IBCEvent> {
-    event_types().get(&event.type_str)?;
+fn extract_attributes_from_tx(event: &tendermint::abci::Event) -> Attributes {
     let mut attr = Attributes::default();
 
-    for tag in event.attributes {
-        match tag.key.as_ref() {
-            CLIENT_ID_ATTRIBUTE_KEY => attr.client_id = tag.value.to_string().parse().unwrap(),
-            CLIENT_TYPE_ATTRIBUTE_KEY => attr.client_type = tag.value.to_string().parse().unwrap(),
-            CONSENSUS_HEIGHT_ATTRIBUTE_KEY => {
-                attr.consensus_height = tag.value.to_string().try_into().unwrap()
-            }
-            _ => {}
+    for tag in &event.attributes {
+        let key = tag.key.as_ref();
+        let value = tag.value.as_ref();
+        match key {
+            CLIENT_ID_ATTRIBUTE_KEY => attr.client_id = value.parse().unwrap(),
+            CLIENT_TYPE_ATTRIBUTE_KEY => attr.client_type = value.parse().unwrap(),
+            CONSENSUS_HEIGHT_ATTRIBUTE_KEY => attr.consensus_height = value.parse().unwrap(),
+            // TODO: `Attributes` has 4 fields and we're only parsing 3
+            _ => panic!("unexpected attribute key: {}", key),
         }
     }
 
-    match event.type_str.as_str() {
-        CREATE_EVENT_TYPE => Some(IBCEvent::CreateClient(CreateClient(attr))),
-        UPDATE_EVENT_TYPE => Some(IBCEvent::UpdateClient(UpdateClient(attr))),
-        _ => None,
-    }
+    attr
 }
 
 /// NewBlock event signals the committing & execution of a new block.
@@ -101,6 +101,12 @@ impl CreateClient {
     }
 }
 
+impl From<Attributes> for CreateClient {
+    fn from(attrs: Attributes) -> Self {
+        CreateClient(attrs)
+    }
+}
+
 impl TryFrom<RawObject> for CreateClient {
     type Error = BoxError;
     fn try_from(obj: RawObject) -> Result<Self, Self::Error> {
@@ -109,7 +115,7 @@ impl TryFrom<RawObject> for CreateClient {
             height: obj.height,
             client_id: attribute!(obj, "create_client.client_id"),
             client_type: attribute!(obj, "create_client.client_type"),
-            consensus_height: consensus_height_str.try_into()?,
+            consensus_height: consensus_height_str.as_str().try_into()?,
         }))
     }
 }
@@ -128,8 +134,15 @@ impl UpdateClient {
     pub fn client_id(&self) -> &ClientId {
         &self.0.client_id
     }
+
     pub fn height(&self) -> &tendermint::block::Height {
         &self.0.height
+    }
+}
+
+impl From<Attributes> for UpdateClient {
+    fn from(attrs: Attributes) -> Self {
+        UpdateClient(attrs)
     }
 }
 
@@ -141,7 +154,7 @@ impl TryFrom<RawObject> for UpdateClient {
             height: obj.height,
             client_id: attribute!(obj, "update_client.client_id"),
             client_type: attribute!(obj, "update_client.client_type"),
-            consensus_height: consensus_height_str.try_into()?,
+            consensus_height: consensus_height_str.as_str().try_into()?,
         }))
     }
 }
@@ -165,7 +178,7 @@ impl TryFrom<RawObject> for ClientMisbehavior {
             height: obj.height,
             client_id: attribute!(obj, "client_misbehaviour.client_id"),
             client_type: attribute!(obj, "client_misbehaviour.client_type"),
-            consensus_height: consensus_height_str.try_into()?,
+            consensus_height: consensus_height_str.as_str().try_into()?,
         }))
     }
 }